Add Action Taken filter to complaint search
Added the Action Taken filter to the Complaint list for CEEB users. The
filter also applies to the map view. In both the complaint search and
map search, if and only if the filters contain an Action Taken, a call
to the Case Management service is made requesting the lead identifiers
(which in NATCom are complaint identifiers) of all leads belonging to
a case_file that has had the specified action taken on it. If the list
is not empty, we add an AND clause to the search to filter the results
against the list. If the list is empty then the search should return
no results, however SQL errors if you pass an empty list to an IN.
To avoid this, a list of a single value that is not a valid complaint
identifier is passed in its stead, giving the desired result.