Add link process to createAssessment for duplicates
When creating an assessment that is marking one complaint as a duplicate
of another, a new linked_complaint_xref row needs to be added, and the
duplicate needs to have its status changed to closed. These take place
in a transaction on the NATCom db. While this transaction is open, the
assessment is sent to CM. If any of this fails, the transaction is rolled
back rather than commit.