Handle closeComplaintOnSuccess parameter in createAssessment
The context in which createAssessment is called dictates whtether
the complaint is closed. The creation of a link and change of status
are independent, but both require a transaction to wrap them with
call to Case Management which needs to happen in both cases. This
handles the conditional logic. It also rejects the promise if the
process fails.