Fix mapSearch agency check for CEEB users
Complaints service's mapSearch function uses the local function _getAgencyByUser
to determine a users agency. This function identifies a users agency through
the office they are assigned to. CEEB users have no office assigned, thus
_getAgencyByUser fails. To get around this, mapSearch now uses the hasCEEBRole
variable already included to determine if _getAgencyByUser should be called, or
"EPO" should be assigned (as is the case for CEEB users). Current implementation
of tests do not address the problem, and there was an existing false positive in
the tests for the complaints service. This will be addressed separately.